--- a/package.json
+++ b/package.json
@@ -32,7 +32,7 @@
 		"package.json"
 	],
 	"scripts": {
-		"build": "tsup",
+		"build": "tsdown",
 		"format": "prettier .",
 		"lint": "eslint . --max-warnings 0",
 		"lint:knip": "knip",
@@ -78,7 +78,7 @@
 		"prettier-plugin-sh": "0.17.0",
 		"release-it": "18.1.2",
 		"sentences-per-line": "0.3.0",
-		"tsup": "8.4.0",
+		"tsdown": "0.9.2",
 		"typescript": "5.8.2",
 		"typescript-eslint": "8.30.0",
 		"vitest": "3.1.1"