--- a/tsconfig.json
+++ b/tsconfig.json
@@ -9,7 +9,8 @@
 		"resolveJsonModule": true,
 		"skipLibCheck": true,
 		"strict": true,
-		"target": "ES2022"
+		"target": "ES2022",
+		"isolatedDeclarations": true
 	},
 	"include": ["src"]
 }