<!-- 👋 Hi, thanks for sending a PR to eslint-fix-utils! 🔧
Please fill out all fields below and make sure each item is true and [x] checked.
Otherwise we may not be able to review your PR. -->

## PR Checklist

- [x] Addresses an existing open issue: fixes #118 
- [x] That issue was marked as [`status: accepting prs`](https://github.com/JoshuaKGoldberg/eslint-fix-utils/issues?q=is%3Aopen+is%3Aissue+label%3A%22status%3A+accepting+prs%22)
- [x] Steps in [CONTRIBUTING.md](https://github.com/JoshuaKGoldberg/eslint-fix-utils/blob/main/.github/CONTRIBUTING.md) were taken

## Overview

This change migrates our build from tsup to tsdown, and enables `isolatedDeclarations` in order to reap all of the performance benefits. You can see it's around 17x faster, even with this tiny project.  

tsup vs tsdown with `isolatedDeclarations`
~1100ms => 84ms

tsdown without and with `isolatedDeclarations`
~1500ms => 84ms

### tsup
![screenshot of build time before](https://github.com/user-attachments/assets/7894161e-10d2-4e92-9932-acea3d530caa)


### tsdown without `isolatedDeclarations`
![screenshot of build time with tsdown and no isolatedDeclarations](https://github.com/user-attachments/assets/fec85a40-b37b-4cae-81e3-f6aab095f6e6)

### tsdown with `isolatedDeclarations`
![screenshot of build time after](https://github.com/user-attachments/assets/00f3e6aa-33de-4adb-94ae-96282352d77a)

So `tsup` seems to be faster if you don't enable `isolatedDeclarations`, but `tsdown` is significantly faster if you do.  The only changes I had to make in order to enable it, was add explicit function return types, which is generally a good practice, for improved type checking performance.